requirements.txt

fastapi==0.75.0
uvicorn==0.17.6
sqlalchemy==1.4.27
python-multipart==0.0.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.7.5
